requirements.txt
playwright
python-dotenv
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
gallery-dl
p115client